% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenCOMET_functions.R
\name{decode_B}
\alias{decode_B}
\title{Decode parameter vector into emission matrix}
\usage{
decode_B(par, observed_states, hidden_states)
}
\arguments{
\item{par}{Parameter vector of length n_obs * n_hidden}

\item{observed_states}{Character vector of observed state names}

\item{hidden_states}{Character vector of hidden state names}
}
\value{
Column-stochastic emission matrix B (n_obs × n_hidden)
}
\description{
Converts unconstrained parameters to column-stochastic emission matrix B.
}
\examples{
par <- rnorm(9)
observed_states <- c("G1", "S", "G2M")
hidden_states <- c("E", "H", "M")
B <- decode_B(par, observed_states, hidden_states)
}
