% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenCOMET_functions.R
\name{proj_columns_simplex}
\alias{proj_columns_simplex}
\title{Project matrix columns onto probability simplex}
\usage{
proj_columns_simplex(B)
}
\arguments{
\item{B}{Matrix whose columns should be projected}
}
\value{
Matrix with columns projected onto simplex
}
\description{
Projects each column of a matrix onto the probability simplex.
}
\examples{
B <- matrix(runif(9), 3, 3)
proj_columns_simplex(B)
}
