% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenCOMET_functions.R
\name{proj_simplex}
\alias{proj_simplex}
\title{Project vector onto probability simplex}
\usage{
proj_simplex(v)
}
\arguments{
\item{v}{Numeric vector to project}
}
\value{
Numeric vector projected onto simplex
}
\description{
Projects a vector onto the probability simplex (sum = 1, all elements >= 0)
using the algorithm from Duchi et al. (2008).
}
\examples{
proj_simplex(c(0.3, 0.8, -0.1))
}
\references{
Duchi, J., Shalev-Shwartz, S., Singer, Y., & Chandra, T. (2008). 
Efficient projections onto the l1-ball for learning in high dimensions. ICML.
}
