% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenCOMET_functions.R
\name{rnd}
\alias{rnd}
\title{Pretty round with trailing zeros}
\usage{
rnd(x, d = 4)
}
\arguments{
\item{x}{Numeric vector to format}

\item{d}{Number of decimal places (default: 4)}
}
\value{
Character vector of formatted numbers
}
\description{
Formats numbers with specified decimal places and trailing zeros.
}
\examples{
rnd(c(0.123456, 0.1, 0.9999), d = 3)
}
