% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenCOMET_functions.R
\name{softmax_n}
\alias{softmax_n}
\title{Softmax function for probability normalization}
\usage{
softmax_n(z)
}
\arguments{
\item{z}{Vector of logits}
}
\value{
Vector of probabilities (sums to 1)
}
\description{
Converts a vector of logits to probabilities using softmax.
}
\examples{
softmax_n(c(1, 2, 3))
}
